requirements.txt
Flask==3.0.2
flask-bcrypt==1.0.1
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
email-validator==2.1.0.post1
python-dotenv==1.0.1
Werkzeug==3.0.1
